requirements.txt
# Core Libraries
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
torch==2.4.0
imbalanced-learn==0.12.3
# gRPC and Protobuf
grpcio==1.65.1
grpcio-tools==1.65.1
protobuf==5.27.3
# Configuration
PyYAML==6.0.1
# Utilities
uuid==1.30

# 工具和辅助库
tqdm==4.65.0
scipy==1.10.1

# 数据处理
joblib==1.2.0 